updating modules
added modules. still needs fixing
