+++ js/modules/imageHover.js
@@ -1,0 +1,17 @@
+// imageHover.js
+import { gsap } from "gsap";
+
+document.addEventListener("mousemove", function(slanting) {
+    var width = window.innerWidth,
+        height = window.innerHeight,
+        positionX = (slanting.clientX/width) - 0.6,
+        positionY = (slanting.clientY/height) - 0.6;
+        gsap.to(".image_con img", {
+    
+            rotationY: positionX * 50,
+            rotationX: -positionY * 50,
+            ease: "none"
+    
+        });
+        
+});
